+++ Source/sync.cpp
@@ -106,7 +106,7 @@ void SyncPlrInv(TSyncHeader *pHdr)
 		auto &item = Items[pHdr->bItemI];
 		pHdr->bItemX = item.position.x;
 		pHdr->bItemY = item.position.y;
-		pHdr->wItemIndx = SDL_SwapLE16(item.IDidx);
+		pHdr->wItemIndx = SDL_SwapLE16(static_cast<int16_t>(item.IDidx));
 		if (item.IDidx == ItemIndex::Ear) {
 			pHdr->wItemCI = SDL_SwapLE16((item._iIName[0] << 8) | item._iIName[1]);
 			pHdr->dwItemSeed = SDL_SwapLE32((item._iIName[2] << 24) | (item._iIName[3] << 16) | (item._iIName[4] << 8) | item._iIName[5]);
@@ -136,7 +136,7 @@ void SyncPlrInv(TSyncHeader *pHdr)
 	const auto &item = MyPlayer->InvBody[sgnSyncPInv];
 	if (!item.isEmpty()) {
 		pHdr->bPInvLoc = sgnSyncPInv;
-		pHdr->wPInvIndx = SDL_SwapLE16(item.IDidx);
+		pHdr->wPInvIndx = SDL_SwapLE16(static_cast<int16_t>(item.IDidx));
 		pHdr->wPInvCI = SDL_SwapLE16(item._iCreateInfo);
 		pHdr->dwPInvSeed = SDL_SwapLE32(item._iSeed);
 		pHdr->bPInvId = item._iIdentified ? 1 : 0;
